% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-add_paths.R
\name{add_paths}
\alias{add_paths}
\title{Add paths to an object of class 'tidy_sem'}
\usage{
add_paths(model, ...)
}
\arguments{
\item{model}{An object of class \code{tidy_sem}.}

\item{...}{Paths to add or substitute, specified in
\code{\link{lavaan}{model.syntax}}, and separated by commas.}
}
\value{
An object of class \code{tidy_sem}.
}
\description{
Add paths to an object of class \code{tidy_sem}, or replace
existing paths. The paths must be specified as
\code{\link{lavaan}{model.syntax}}, and separated by commas.
}
\details{
Currently, only the \code{\link{lavaan}{lavaan}} commands \code{~,
~~, =~,} and \code{~1} are parsed.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lavaan]{model.syntax}}
}
