% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{all_fun}
\alias{all_fun}
\alias{all_fun.sem_graph}
\alias{all_sig}
\alias{hide_sig}
\alias{show_sig}
\alias{color_sig}
\alias{colour_sig}
\alias{linetype_sig}
\alias{size_sig}
\alias{alpha_sig}
\alias{all_non_sig}
\alias{hide_non_sig}
\alias{show_non_sig}
\alias{color_non_sig}
\alias{colour_non_sig}
\alias{linetype_non_sig}
\alias{size_non_sig}
\alias{alpha_non_sig}
\alias{all_fixed}
\alias{hide_fixed}
\alias{show_fixed}
\alias{color_fixed}
\alias{colour_fixed}
\alias{linetype_fixed}
\alias{size_fixed}
\alias{alpha_fixed}
\alias{all_pos}
\alias{hide_pos}
\alias{show_pos}
\alias{color_pos}
\alias{colour_pos}
\alias{linetype_pos}
\alias{size_pos}
\alias{alpha_pos}
\alias{all_neg}
\alias{hide_neg}
\alias{show_neg}
\alias{color_neg}
\alias{colour_neg}
\alias{linetype_neg}
\alias{size_neg}
\alias{alpha_neg}
\alias{all_var}
\alias{hide_var}
\alias{show_var}
\alias{color_var}
\alias{colour_var}
\alias{linetype_var}
\alias{size_var}
\alias{alpha_var}
\alias{all_cov}
\alias{hide_cov}
\alias{show_cov}
\alias{color_cov}
\alias{colour_cov}
\alias{linetype_cov}
\alias{size_cov}
\alias{alpha_cov}
\alias{all_reg}
\alias{hide_reg}
\alias{show_reg}
\alias{color_reg}
\alias{colour_reg}
\alias{linetype_reg}
\alias{size_reg}
\alias{alpha_reg}
\alias{all_loading}
\alias{hide_loading}
\alias{show_loading}
\alias{color_loading}
\alias{colour_loading}
\alias{linetype_loading}
\alias{size_loading}
\alias{alpha_loading}
\alias{all_fixed_loading}
\alias{hide_fixed_loading}
\alias{show_fixed_loading}
\alias{color_fixed_loading}
\alias{colour_fixed_loading}
\alias{linetype_fixed_loading}
\alias{size_fixed_loading}
\alias{alpha_fixed_loading}
\alias{all_mean}
\alias{hide_mean}
\alias{show_mean}
\alias{color_mean}
\alias{colour_mean}
\alias{linetype_mean}
\alias{size_mean}
\alias{alpha_mean}
\title{Apply an expression to an object based on a condition}
\usage{
all_fun(data, expr, condition, ...)

\method{all_fun}{sem_graph}(data, expr, condition, element = "edges", ...)

all_sig(data, expr, ...)

hide_sig(data, ...)

show_sig(data, ...)

color_sig(data, color = "black", ...)

colour_sig(data, colour = "black", ...)

color_sig(data, color = "black", ...)

linetype_sig(data, linetype = 1, ...)

size_sig(data, size = 1, ...)

alpha_sig(data, alpha = 1, ...)

all_non_sig(data, expr, ...)

hide_non_sig(data, ...)

show_non_sig(data, ...)

color_non_sig(data, color = "black", ...)

colour_non_sig(data, colour = "black", ...)

color_non_sig(data, color = "black", ...)

linetype_non_sig(data, linetype = 1, ...)

size_non_sig(data, size = 1, ...)

alpha_non_sig(data, alpha = 1, ...)

all_fixed(data, expr, ...)

hide_fixed(data, ...)

show_fixed(data, ...)

color_fixed(data, color = "black", ...)

colour_fixed(data, colour = "black", ...)

color_fixed(data, color = "black", ...)

linetype_fixed(data, linetype = 1, ...)

size_fixed(data, size = 1, ...)

alpha_fixed(data, alpha = 1, ...)

all_pos(data, expr, ...)

hide_pos(data, ...)

show_pos(data, ...)

color_pos(data, color = "black", ...)

colour_pos(data, colour = "black", ...)

color_pos(data, color = "black", ...)

linetype_pos(data, linetype = 1, ...)

size_pos(data, size = 1, ...)

alpha_pos(data, alpha = 1, ...)

all_neg(data, expr, ...)

hide_neg(data, ...)

show_neg(data, ...)

color_neg(data, color = "black", ...)

colour_neg(data, colour = "black", ...)

color_neg(data, color = "black", ...)

linetype_neg(data, linetype = 1, ...)

size_neg(data, size = 1, ...)

alpha_neg(data, alpha = 1, ...)

all_var(data, expr, ...)

hide_var(data, ...)

show_var(data, ...)

color_var(data, color = "black", ...)

colour_var(data, colour = "black", ...)

color_var(data, color = "black", ...)

linetype_var(data, linetype = 1, ...)

size_var(data, size = 1, ...)

alpha_var(data, alpha = 1, ...)

all_cov(data, expr, ...)

hide_cov(data, ...)

show_cov(data, ...)

color_cov(data, color = "black", ...)

colour_cov(data, colour = "black", ...)

color_cov(data, color = "black", ...)

linetype_cov(data, linetype = 1, ...)

size_cov(data, size = 1, ...)

alpha_cov(data, alpha = 1, ...)

all_reg(data, expr, ...)

hide_reg(data, ...)

show_reg(data, ...)

color_reg(data, color = "black", ...)

colour_reg(data, colour = "black", ...)

color_reg(data, color = "black", ...)

linetype_reg(data, linetype = 1, ...)

size_reg(data, size = 1, ...)

alpha_reg(data, alpha = 1, ...)

all_loading(data, expr, ...)

hide_loading(data, ...)

show_loading(data, ...)

color_loading(data, color = "black", ...)

colour_loading(data, colour = "black", ...)

color_loading(data, color = "black", ...)

linetype_loading(data, linetype = 1, ...)

size_loading(data, size = 1, ...)

alpha_loading(data, alpha = 1, ...)

all_fixed_loading(data, expr, ...)

hide_fixed_loading(data, ...)

show_fixed_loading(data, ...)

color_fixed_loading(data, color = "black", ...)

colour_fixed_loading(data, colour = "black", ...)

color_fixed_loading(data, color = "black", ...)

linetype_fixed_loading(data, linetype = 1, ...)

size_fixed_loading(data, size = 1, ...)

alpha_fixed_loading(data, alpha = 1, ...)

all_mean(data, expr, ...)

hide_mean(data, ...)

show_mean(data, ...)

color_mean(data, color = "black", ...)

colour_mean(data, colour = "black", ...)

color_mean(data, color = "black", ...)

linetype_mean(data, linetype = 1, ...)

size_mean(data, size = 1, ...)

alpha_mean(data, alpha = 1, ...)
}
\arguments{
\item{data}{Object to manipulate.}

\item{expr}{Expression to perform on elements of \code{data} for which
\code{condition == TRUE}.}

\item{condition}{Expression that evaluates to a logical vector in the
environment of \code{data}.}

\item{...}{Additional arguments passed to and from functions.}

\item{element}{Character vector. The elements of the \code{sem_graph} to
edit, defaults to \code{"edges"}.}

\item{color}{Atomic character vector, indicating which color to assign to
the selected elements.}

\item{colour}{Atomic character vector, indicating which colour to assign to
the selected elements.}

\item{linetype}{Atomic character vector, indicating which linetype to assign to
the selected elements.}

\item{size}{Atomic character vector, indicating which size to assign to
the selected elements.}

\item{alpha}{Atomic character vector, indicating which alpha to assign to
the selected elements.}
}
\value{
Object of the same class as \code{data}.
}
\description{
This function allows users to conditionally manipulate an
object.
}
\examples{
library(lavaan)
res <- sem("dist ~ speed", cars, meanstructure = TRUE)
all_fun(table_results(res), {label = "bla"}, {pval < .05})
}
