% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-as_syntax.R
\name{as.lavaan}
\alias{as.lavaan}
\title{Convert tidy_sem to lavaan syntax}
\usage{
as.lavaan(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidy_sem}}

\item{...}{Additional parameters to be passed to and from functions.}
}
\value{
Character vector.
}
\description{
Final stage in the tidySEM workflow for syntax generation:
Convert the tidy_sem object to lavaan syntax.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
