% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{as.layout}
\alias{as.layout}
\title{Convert object to layout}
\usage{
as.layout(x)
}
\arguments{
\item{x}{Object to convert to a tidy_layout. The default argument reads a
selected matrix from the clipboard.
To use this functionality, specify your layout in a spreadsheet program,
select the block of cells, and copy it to the clipboard.}
}
\value{
Object of class 'tidy_layout'
}
\description{
Convert an object to a tidy_layout for a SEM graph.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{tidy_graph}
