% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-as_syntax.R
\name{as_lavaan}
\alias{as_lavaan}
\title{Convert tidy_sem to 'lavaan' syntax}
\usage{
as_lavaan(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidy_sem}}

\item{...}{Additional parameters to be passed to and from functions.}
}
\value{
Character vector.
}
\description{
Final stage in the 'tidySEM' workflow for syntax generation:
Convert the \code{tidy_sem} object to \code{lavaan} syntax in tabular format
(see \code{\link[lavaan]{model.syntax}}).
}
\examples{
mod <- list(syntax = structure(list(lhs = "x", op = "~", rhs = "y",
                                    free = TRUE, value = "", label = "",
                                    category = "", aspect = ""),
            class = "data.frame", row.names = c(NA, -1L)))
class(mod) <- "tidy_sem"
as_lavaan(mod)
}
