% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-as_syntax.R
\name{as_mplus}
\alias{as_mplus}
\title{Convert tidy_sem to Mplus syntax}
\usage{
as_mplus(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidy_sem}.}

\item{...}{Additional parameters to be passed to and from functions.}
}
\value{
Character vector.
}
\description{
Final stage in the tidySEM workflow for syntax generation:
Convert the tidy_sem object to Mplus syntax.
}
\examples{
mod <- list(syntax = tidySEM:::check_lav_tab(data.frame(
lhs = "x", op = "~", rhs = "y", free = TRUE)))
class(mod) <- "tidy_sem"
as_mplus(mod)
}
