% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-3step.R
\name{bch_test}
\alias{bch_test}
\title{Extract BCH test from object}
\usage{
bch_test(x, ...)
}
\arguments{
\item{x}{An object with attribute `test`.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
An object of class `bch_test` and `list`.
}
\description{
For a model generated by [BCH()], extract equivalence tests.
}
\examples{
a <- "test"
attr(a, "test") <- "The test"
bch_test(a)
}
