% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-bifactor.R
\name{bifactor}
\alias{bifactor}
\title{Generate syntax for a bifactor model}
\usage{
bifactor(x, ...)
}
\arguments{
\item{x}{An object for which a method exists, including \code{tidy_sem}
(generated using \code{\link[tidySEM]{dictionary}}, or \code{data.frame} (for
which \code{\link[tidySEM]{dictionary}} will be run first).}

\item{...}{Additional parameters passed to \code{\link{measurement}}.}
}
\value{
An object of class \code{tidy_sem_bifactor} and \code{tidy_sem} which can be passed to \code{\link{run_bifactor}}.
}
\description{
Generate syntax for a bifactor model for latent variables.
}
\examples{
dict <- tidy_sem(c("a_1", "a_2", "a_3", "b_4", "b_5"))
bifactor(dict)
}
