% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrt.R
\name{BLRT}
\alias{BLRT}
\title{Conduct Bootstrapped Likelihood Ratio Test}
\usage{
BLRT(x, replications = 100, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{replications}{Integer reflecting the number of bootstrapped
replications, defaults to \code{100}.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A data.frame.
}
\description{
Conduct Bootstrapped Likelihood Ratio Test to compare two
mixture models.
}
\examples{
\dontrun{
df <- iris[, 1, drop = FALSE]
names(df) <- "x"
res <- mx_mixture(model = "x ~ m{C}*1
                           x ~~ v{C}*x", classes = 1:2, data = df)
BLRT(res, replications = 4)
}
}
