% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_bootstrap.R
\name{chi_sq}
\alias{chi_sq}
\title{Calculate Chi Square Statistic}
\usage{
chi_sq(x, y)
}
\arguments{
\item{x}{An object for which a method exists
(e.g, \code{data.frame}).}

\item{y}{An object for which a method exists
(e.g, \code{data.frame}).}
}
\value{
\code{numeric}
}
\description{
Given two datasets with ordinal variables,
computes the lambda statistic, which is the chi-square statistic minus the
degrees of freedom.
}
\examples{
\dontrun{
if(interactive()){
 chi_sq(iris[1:2], iris[3:4])
}
}
}
