% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scales.R
\name{create_scales}
\alias{create_scales}
\title{Create scale scores from observed variables}
\usage{
create_scales(
  data,
  keys.list,
  missing = TRUE,
  impute = "none",
  omega = NULL,
  write_files = FALSE,
  digits = 2,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing all variables referenced in the
\code{keys.list}.}

\item{keys.list}{A named list, indicating which variables belong to which
scale. See \code{\link[psych]{scoreItems}} and
\code{\link[psych]{make.keys}} for more information.}

\item{missing}{Whether to use rows with partially missing values, Default: TRUE.
See: \code{\link[psych]{scoreItems}}.}

\item{impute}{Method for handling missing values, Default: 'none'. This
default method uses all available data to calculate scale scores, which is
acceptable for mean scales, but not for sum scales. See:
\code{\link[psych]{scoreItems}}.}

\item{omega}{Which of McDonald's \code{\link[psych]{omega}} coefficients to
report. Default: NULL; valid options include: \code{"omega_h"},
\code{"omega.lim"}, \code{"alpha"}, \code{"omega.tot"}, \code{"G6"}.}

\item{write_files}{Whether to write the scale descriptive and correlation
tables to .csv files, Default: FALSE}

\item{digits}{Number of digits for rounding, Default: 2}

\item{...}{Additional parameters to pass to and from functions.}
}
\value{
List with elements: \code{$descriptives}, \code{$correlations}, and
\code{$scores}.
}
\description{
This function calculates mean or sum scores from a
\code{data.frame} and a named list describing the items in each scale. It
returns the scores, a scale descriptive table, and a scale correlation table.
It relies on several functions from the
\code{psych} package; most notably \code{\link[psych]{scoreItems}},
\code{\link[psych]{describe}}, \code{\link[psych]{fa}}, and
\code{\link[psych]{omega}}.
}
\examples{
out <- create_scales(iris, keys.list = list(scalename = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")))
out$descriptives
}
