% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data_mix_ordinal.R
\docType{data}
\name{data_mix_ordinal}
\alias{data_mix_ordinal}
\title{Simulated data for mixture model with ordinal indicators}
\format{
A data frame with 5000 rows and 4 variables.
}
\usage{
data(data_mix_ordinal)
}
\description{
This simulated dataset, based on the 'Mplus' User's Guide example 7.6,
contains four columns of integer data that should be treated as ordinal.
}
\details{
\tabular{lll}{
\strong{u1} \tab \code{integer} \tab Indicator 1, should be treated as
ordinal.\cr
\strong{u2} \tab \code{integer} \tab Indicator 2, should be treated as
ordinal.\cr
\strong{u3} \tab \code{integer} \tab Indicator 3, should be treated as
ordinal.\cr
\strong{u4} \tab \code{integer} \tab Indicator 4, should be treated as
ordinal.
}
}
\references{
Muthén, L.K. and Muthén, B.O. (1998-2017). Mplus User's Guide.
Eighth Edition. Los Angeles, CA: Muthén & Muthén
}
\keyword{datasets}
