% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem_attributes.R
\name{edges}
\alias{edges}
\alias{edges<-.sem_graph}
\title{Extract edges from sem_graph}
\usage{
edges(x)

\method{edges}{sem_graph}(x) <- value
}
\arguments{
\item{x}{Object of class sem_graph.}
}
\value{
data.frame
}
\description{
Provides access to the \code{edges} element of a
\code{sem_graph} object. This can be used to return or assign to the
\code{edges} element.
}
