% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table_results.R
\name{est_sig}
\alias{est_sig}
\title{Add significance asterisks to object}
\usage{
est_sig(x, digits = 2, sig = NULL)
}
\arguments{
\item{x}{An object for which a method exists. This will be treated as numeric
by the default method.}

\item{digits}{Integer. The number of digits to round the estimate column to.}

\item{sig}{Optional, a vector of p-values for the default method.}
}
\value{
A character vector of formatted estimates.
}
\description{
Takes an object, and adds significance asterisks.
}
\examples{
est_sig(c(.222, .3333), sig = c(.054, .045))
}
\seealso{
table_results

Other Reporting tools:
\code{\link[=conf_int]{conf_int()}},
\code{\link[=table_fit]{table_fit()}},
\code{\link[=table_prob]{table_prob()}},
\code{\link[=table_results]{table_results()}}
}
\author{
Caspar J. van Lissa
}
\concept{Reporting tools}
