% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-estimate.R
\name{estimate_mx}
\alias{estimate_mx}
\title{Estimate tidy_sem using 'OpenMx'}
\usage{
estimate_mx(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidy_sem}.}

\item{...}{Additional parameters passed to the estimating function.}
}
\value{
An object of class \code{MxModel}.
}
\description{
This function is a wrapper for the \code{\link{as_ram}}
and \code{\link{run_mx}} functions.
}
\examples{
df <- iris[1:4]
names(df) <- paste0("x_", 1:4)
model <- tidy_sem(df)
model <- measurement(model)
res <- estimate_mx(model)
summary(res)
}
