% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{filter_graph}
\alias{filter_graph}
\title{Filter graph elements}
\usage{
filter_graph(x, subset, select, element = "edges", ...)
}
\arguments{
\item{x}{An object of class \code{sem_graph}.}

\item{subset}{logical expression indicating elements or rows to keep:
missing values are taken as false.}

\item{select}{Expression, indicating columns to select from a data frame.}

\item{element}{Character. The element of the \code{sem_graph} to filter,
defaults to \code{"edges"}.}

\item{...}{Arguments passed on to \code{\link[base]{within}}.}
}
\value{
An object of class \code{sem_graph}.
}
\description{
Apply a filter to one of the elements
of a \code{sem_graph} object, and return the modified \code{sem_graph}. Uses
the function \code{\link[base]{subset}}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
