% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_layout.R
\name{generate_layout}
\alias{generate_layout}
\title{Generate a rudimentary layout from a model object}
\usage{
generate_layout(x, layout_algorithm = "layout.reingold.tilford")
}
\arguments{
\item{x}{A model for which a method exists.}

\item{layout_algorithm}{Which algorithm to use, from the
\code{\link[igraph]{igraph-package}}. The default Reingold-Tilford algorithm
is most suitable for SEM graphs.}
}
\value{
Matrix
}
\description{
This is a wrapper function to the
\code{\link[igraph]{layout.reingold.tilford}} function, or other layout
functions from the \code{\link[igraph]{igraph-package}}. It returns a layout
in matrix format.
}
\examples{
\dontrun{
library(lavaan)
fit <- sem("Sepal.Length ~ Petal.Width", data = iris)
generate_layout(fit)
}
}
