% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from model object}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Additional arguments passed to and from functions.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts data from an object for which a method exists.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
