% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_attributes.R
\name{get_dict}
\alias{get_dict}
\alias{get_dict<-}
\title{Extract data_dict from sem_syntax}
\usage{
get_dict(x)

get_dict(x) <- value
}
\arguments{
\item{x}{Object of class sem_syntax.}

\item{value}{A valid value for \code{get_dict(x)}.}
}
\value{
data.frame
}
\description{
Provides access to the \code{data_dict} element of a
\code{sem_syntax} object. This can be used to return or assign to the
\code{data_dict} element.
}
