% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test.R
\name{get_estimates}
\alias{get_estimates}
\title{Get estimates from a model object}
\usage{
get_estimates(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Parameters passed to and from other functions.}
}
\value{
An object of class 'model_estimates'
}
\description{
Get estimates from a model object.
This convenience function allows you to see that coefficients are properly
extracted, note how their names will be parsed, and inspect their values.
}
\keyword{internal}
