% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_summaries.R
\name{get_fit}
\alias{get_fit}
\title{Get fit indices from objects}
\usage{
get_fit(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A data.frame.
}
\description{
Get fit indices from objects for which a method exists.
}
\examples{
\dontrun{
df <- iris[, 1, drop = FALSE]
names(df) <- "x"
res <- mx_mixture(model = "x ~ m{C}*1
                           x ~~ v{C}*x", classes = 1, data = df)
table_fit(res)
}
}
