% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{get_nodes}
\alias{get_nodes}
\title{Extract nodes from a SEM model object}
\usage{
get_nodes(x, ...)
}
\arguments{
\item{x}{A model object of class 'mplusObject' or 'lavaan'.}

\item{...}{Additional parameters to be passed to and from other functions.}
}
\value{
An object of class 'tidy_nodes'
}
\description{
Attempts to extract nodes from a SEM model object, where nodes
are defined as observed or latent variables.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{tidy_graph}
