% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{graph}
\alias{graph}
\title{Render a graph}
\usage{
graph(model, layout, rect_width = 1.2, rect_height = 0.8,
  ellipses_a = 1, ellipses_b = 1, spacing_x = 2, spacing_y = 2,
  text_size = 4, curvature = 0.1, angle = NULL)
}
\arguments{
\item{model}{A model object for which a method exists (e.g.,
\code{mplus.model} or \code{lavaan}).}

\item{layout}{An object of class \code{tidy_layout}, or a matrix with the
layout of the graph that can be converted using
\code{\link[tidySEM]{get_layout}}.}

\item{rect_width}{Width of rectangles (used to display observed variables),
Default: 1.2}

\item{rect_height}{Height of rectangles (used to display observed variables),
Default: 0.8}

\item{ellipses_a}{Width of ellipses (used to display latent variables),
Default: 1}

\item{ellipses_b}{Height of ellipses (used to display latent variables),
Default: 1}

\item{spacing_x}{Spacing between columns of the graph, Default: 1}

\item{spacing_y}{Spacing between rows of the graph, Default: 1}

\item{text_size}{Point size of text, Default: 4}

\item{curvature}{Curvature of curved connectors. To flip connectors, use
negative values. Default: .1}

\item{angle}{Angle used to connect nodes by the top and bottom. Defaults to
NULL, which means Euclidean distance is used to determine the shortest
distance between node sides. A numeric value between 0-180 can be provided,
where 0 means that only nodes with the same x-coordinates are connected
top-to-bottom, and 180 means that all nodes are connected top-to-bottom.}
}
\value{
Object of class 'sem_graph'
}
\description{
Render a graph based on a model and layout.
}
\details{
Calls the functions \code{\link[tidySEM]{get_nodes}} and
\code{\link[tidySEM]{get_edges}} on \code{model}, before calling
\code{\link[tidySEM]{prepare_graph}} and \code{plot}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{tidy_graph}
