% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{get_layout}
\alias{get_layout}
\title{Generate graph layout}
\usage{
get_layout(..., rows = NULL)
}
\arguments{
\item{...}{Character arguments corresponding to layout elements. Use node
names, empty strings (""), or NA values.}

\item{rows}{Numeric, indicating the number of rows of the graph.}
}
\value{
Object of class 'tidy_layout'
}
\description{
Generate a tidy_layout for a SEM graph by specifying node names,
and empty strings or \code{NA} values for spaces.
}
\examples{
get_layout("c", "",  "d",
           "",  "e", "", rows = 2)
}
\seealso{
as.layout
}
\keyword{tidy_graph}
