% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lcs}
\alias{lcs}
\title{Longest common substring}
\usage{
lcs(x, y)
}
\arguments{
\item{x}{First atomic character vector.}

\item{y}{Second atomic character vector.}
}
\value{
Character string
}
\description{
Extract the longest common substring of two strings using
base R.
}
\examples{
lcs("test1", "test2")
}
\keyword{mplus}
\keyword{utilities}
