% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmx_functions.R
\name{lsub}
\alias{lsub}
\alias{as_ram}
\title{Apply pattern replacement over a vector}
\usage{
lsub(x, replacement = NULL, pattern = "{C}", fixed = TRUE, ...)

as_ram(x, ...)
}
\arguments{
\item{x}{A character vector, describing the user-specified model using
the lavaan model syntax.}

\item{replacement}{a character vector of length 1 or more. Each element is
applied to \code{x} in turn. Default: NULL}

\item{pattern}{A character string containing a regular expression (or
character string when \code{fixed = TRUE}). Default: \code{'{C}'}.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.
Default: TRUE}

\item{...}{Parameters passed on to other functions.}
}
\value{
A list of results returned by \code{\link[base:grep]{gsub}}.

Returns a new MxModel object \code{\link[OpenMx]{mxModel}}.
}
\description{
\code{lsub} returns a list of the same length as
\code{replacement}, each element of which is the result of applying
\code{\link[base:grep]{gsub}} to \code{x} using \code{\link{lapply}}.

Converts models specified using lavaan syntax to RAM models for
\code{OpenMx}.
}
\details{
The procedure is as follows:
\enumerate{
 \item Apply \code{\link[lavaan]{lavaanify}} to the \code{model}. The default
 arguments to \code{\link[lavaan]{lavaanify}} correspond to those of the
 \code{\link[lavaan]{sem}} function.
 \item Convert each row of the resulting lavaan parameter table to a
 \code{\link[OpenMx]{mxPath}}.
 \item Apply \code{\link[OpenMx]{mxModel}} to the \code{mxPath}s to create
 an \code{OpenMx} model using RAM specification
}
}
\examples{
lsub("a{C}", 1:3)
lsub("a{C}", 1:3)
}
