% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motley_functions.R
\name{makeKeysList}
\alias{makeKeysList}
\title{Make a keys list}
\usage{
makeKeysList(var_names, questionnaire_filter = "^\\\\w+\\\\d+$",
  item_filter = "\\\\d+$", skip_questionnaires = NULL,
  questionnaire_length = 3L)
}
\arguments{
\item{var_names}{Character vector of variable names to convert to a KeysList.}

\item{questionnaire_filter}{Character. A regex filter that identifies
variable names that are part of a questionnaire.}

\item{item_filter}{Character. A regex filter that identifies the incremental
part of multiple items belonging to the same questionnaire (e.g., if items
are labelled c("a1", "a2", "a3"), then the item_filter would be \code{"\\d$"}
).}

\item{skip_questionnaires}{Character vector. Names of the questionnaires to
skip.}

\item{questionnaire_length}{Integer. Minimum number of items required to be
included in the list of questionnaires.}
}
\value{
A named list with elements corresponding to the questionnaires
identified. Each element is a character vector with the variable names
belonging to that questionnaire.
}
\description{
Attempts to generate a keys.list, as used by functions from the \code{psych}
package, and several functions in this \code{motley} package.
}
\examples{
makeKeysList(var_names = c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5",
"mac_q_j_1", "mac_q_j_2", "mac_q_j_3", "mac_q_j_4", "mac_q_j_5", "mac_q_j_6",
"mac_q_j_7", "mac_q_j_8", "mac_q_j_9", "mac_q_j_10", "mac_q_j_11",
"mac_q_j_12", "mac_q_j_13", "mac_q_j_14", "mac_q_j_15", "mac_q_j_16",
"mac_q_j_17", "mac_q_j_18", "mac_q_j_19", "mac_q_j_20", "mac_q_j_21",
"mac_q_r_1", "mac_q_r_2", "mac_q_r_3", "mac_q_r_4", "mac_q_r_5", "mac_q_r_6",
"mac_q_r_7", "mac_q_r_8", "mac_q_r_9", "mac_q_r_10", "mac_q_r_11",
"mac_q_r_12", "mac_q_r_13", "mac_q_r_14", "mac_q_r_15", "mac_q_r_16",
"mac_q_r_17", "mac_q_r_18", "mac_q_r_19", "mac_q_r_20", "mac_q_r_21"),
questionnaire_filter = "^[a-zA-Z_]+")
}
\author{
Caspar J. van Lissa
}
