% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_cfa.R
\name{measurement}
\alias{measurement}
\title{Generate syntax for a measurement model}
\usage{
measurement(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{An object for which a method exists, including \code{data_dict}
(generated using \code{\link[tidySEM]{dictionary}}, or \code{data.frame} (for
which \code{\link[tidySEM]{dictionary}} will be run first).}

\item{center}{Whether or not to mean-center the latent variables,
Default: TRUE}

\item{scale}{Whether to identify the model by fixing latent variable variance
to 1 (\code{scale = TRUE}), or by fixing the first factor loading to 1
(\code{scale = FALSE}).}
}
\value{
An object of class \code{sem_syntax}.
}
\description{
Generate syntax for a measurement model for latent variables.
}
\examples{
dict <- get_dictionary(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5"))
measurement(dict)
}
