% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_simple_thresholds.R
\name{mx_deviances}
\alias{mx_deviances}
\title{Create Thresholds Based on Deviances}
\usage{
mx_deviances(
  vars,
  nThresh = NA,
  free = NULL,
  values = NULL,
  labels = NA,
  lbound = NULL,
  ubound = NULL
)
}
\arguments{
\item{vars}{Character vector. Names of variables for which thresholds should
be specified.}

\item{nThresh}{Numeric vector. Number of thresholds for each of variable in
\code{vars}.}

\item{free}{Logical vector, used to fill the \code{free} matrix which indicates
whether the deviance will be freely estimated or fixed.}

\item{values}{Numeric vector of starting values. Default uses
\code{mxNormalQuantiles()}.}

\item{labels}{Character vector. Note that constraints specified using these
labels are applied using \code{mxConstraint()}, because the thresholds are derived
quantities, not parameters of the model (the deviances are).}

\item{lbound}{PARAM_DESCRIPTION, Default: NULL}

\item{ubound}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
A list of \code{mxMatrix} objects, to be included in an \code{mxModel}.
}
\description{
This function creates a list of mxMatrices to specify thresholds
for ordinal variables as a function of strictly positive deviances, see
Details.
}
\details{
Specifying thresholds as a function of strictly positive deviances
is an efficient way to get models with ordinal indicators to converge, if
they tend to arrive at impossible solutions where the thresholds are not
monotonously increasing (e.g., in mixture models). The parameters of a model
specified this way are deviances (i.e., the differences between threshold
values), which have a lower bound of a small positive number to ensure that
thresholds are always increasing. The thresholds are computed by adding
subsequent deviances.
}
