% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_dummies.R
\name{mx_dummies}
\alias{mx_dummies}
\title{Dummy Code Factor Variables}
\usage{
mx_dummies(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Arguments}
}
\value{
A `data.frame`.
}
\description{
For each variable *v* that inherits
`factor`, create a number of new variables equal
to `levels(v)` to indicate group membership (1)
or non-membership (0) of that level.
The resulting dummies have class `mxFactor`.
}
\examples{
mx_dummies(iris[1:5,])
}
