% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_mixed-data.R
\name{mx_mixed_lca}
\alias{mx_mixed_lca}
\title{Estimate mixed data latent class analysis using OpenMx}
\usage{
mx_mixed_lca(data = NULL, classes = 1L, run = TRUE, ...)
}
\arguments{
\item{data}{The data.frame to be used for model fitting.}

\item{classes}{A vector of integers, indicating which class solutions to
generate. Defaults to 1L. E.g., \code{classes = 1:6},}

\item{run}{Logical, whether or not to run the model. This should usually be
set to \code{TRUE}, because this function runs several models to aid in specifying
good starting values; see Details.}

\item{...}{Additional arguments, passed to functions.}
}
\value{
Returns an \code{\link[OpenMx:mxModel]{OpenMx::mxModel()}}.
}
\description{
This function simplifies the specification of latent class models with mixed
data types:
models that estimate membership of a categorical latent variable based on
binary/ordinal and continuous indicators. See Details for more information.
}
\details{
The procedure is as follows:
\enumerate{
\item Estimate a latent profile analysis for the continuous indicators using
\code{\link[mx_profiles]{mx_profiles()}}. Additional arguments, like
\code{variabces = "free"}, can be passed via \code{...}. The estimator uses
simulated annealing.
\item To obtain good starting values for the categorical indicators, use the
classes from step 1. to estimate an auxiliary model for the ordinal
indicators with \code{\link[BCH]{BCH()}}.
\item Estimate a latent class analysis for the categorical indicators using
\code{\link[mx_lca]{mx_lca()}}, with the results of step 2. as starting
values. The estimator uses
\code{\link[OpenMx:mxTryHardOrdinal]{OpenMx::mxTryHardOrdinal()}}.
\item Combine the models from steps 1. and 3. into one joint model. Conduct one
final optimization step using
\code{\link[OpenMx:mxTryHardOrdinal]{OpenMx::mxTryHardOrdinal()}}.
}
}
\examples{
\dontrun{
library(OpenMx)
set.seed(1)
n = 200
mns <- c(rep(0, floor(.3*n)), rep(2, ceiling(.7*n)))
df <- rnorm(4*n, mean = rep(mns, 4))
df <- matrix(df, nrow = n)
df <- t(t(df) * c(1, 2, .5, 1))
df <- data.frame(df)
df$X4 <- cut(df$X4, 3, labels = FALSE)
df$X4 <- OpenMx::mxFactor(df$X4, levels = c(1:3))
res <- mx_mixed_lca(data = df, classes = 2)
}
}
\references{
Van Lissa, C. J., Garnier-Villarreal, M., & Anadria, D. (2023).
Recommended Practices in Latent Class Analysis using the Open-Source
R-Package tidySEM. Structural Equation Modeling.
\doi{10.1080/10705511.2023.2250920}
}
\keyword{mixture}
\keyword{models}
\keyword{openmx}
