% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem_attributes.R
\name{nodes}
\alias{nodes}
\alias{nodes<-.sem_graph}
\title{Extract nodes from sem_graph}
\usage{
nodes(x)

\method{nodes}{sem_graph}(x) <- value
}
\arguments{
\item{x}{Object of class sem_graph.}
}
\value{
data.frame
}
\description{
Provides access to the \code{nodes} element of a
\code{sem_graph} object. This can be used to return or assign to the
\code{nodes} element.
}
