% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-bifactor.R
\name{omega.MxRAMModel}
\alias{omega.MxRAMModel}
\title{Compute omega (openMx)}
\usage{
\method{omega}{MxRAMModel}(
  f,
  total_variance_calculation = c("observed", "modelled"),
  generalFactorName = "G"
)
}
\description{
Compute omega based on a openMx model
}
