% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-bifactor.R
\name{omega}
\alias{omega}
\title{Compute omega}
\usage{
omega(
  f,
  total_variance_calculation = c("observed", "modelled"),
  standardization_method = c("std.all", "std.lv"),
  generalFactorName = "G"
)
}
\arguments{
\item{f}{A model fit object
(fit object generated using \code{\link[tidySEM]{bifactor}}).}

\item{total_variance_calculation}{whether to compute total variances with the
model implied or the observed correlation matrix.
Default: "observed"}

\item{standardization_method}{whether to base omega on the completely standardized
solution or the solution in which latent variables are standardized (see lavaan)
Default: "std.all"}
}
\value{
A dataframe containing omega statistics for the general factor
and the subfactors
(\code{omega} and \code{omegaS}, \code{omegaH} and \code{omegaHS}, and \code{ECV}).
}
\description{
Compute omega based on a model fit object
}
\references{
10.1080/00223891.2015.1089249
10.1080/00223891.2015.1117928
}
