% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_predict.R
\name{predict_class}
\alias{predict_class}
\title{Predict Class Membership}
\usage{
predict_class(object, newdata, ...)
}
\arguments{
\item{object}{An \code{MxModel} object with attribute \code{tidySEM = "mixture"}.}

\item{newdata}{A \code{data.frame} with the same column names and data types as
the one used to estimate the model.}

\item{...}{Other arguments passed to \code{predict.MxModel()}.}
}
\value{
A \code{data.frame}.
}
\description{
Predict class membership for latent class analyses conducted
with \code{tidySEM} in \code{OpenMx}.
}
