% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{prepare_graph}
\alias{prepare_graph}
\title{Prepare graph data}
\usage{
prepare_graph(x, ...)
}
\arguments{
\item{x}{A model object of class 'mplusObject' or 'lavaan'.}

\item{...}{Additional parameters to be passed to and from other functions.}
}
\value{
An object of class 'graph_data'
}
\description{
Extracts nodes and edges from a SEM model object, where nodes
are defined as observed or latent variables, and edges
are defined as regression paths and covariances between variables (nodes).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{tidy_graph}
