% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{prepare_sem_graph}
\alias{prepare_sem_graph}
\title{Prepare to plot a SEM graph}
\usage{
prepare_sem_graph(layout, edges = NULL, nodes = NULL,
  rect_width = 1.2, rect_height = 0.8, ellipses_a = 1,
  ellipses_b = 1, spacing_x = 2, spacing_y = 2, text_size = 4,
  curvature = 0.1, angle = NULL)
}
\arguments{
\item{layout}{A matrix with the layout of the graph, using the same names
for nodes as in the \code{edges} argument, or an object of class
'tidy_layout', created with the \code{\link[tidySEM]{get_layout}} function.}

\item{edges}{Object of class 'tidy_edges', Default: NULL}

\item{nodes}{Object of class 'tidy_nodes', created with the
\code{\link[tidySEM]{get_nodes}} function.}

\item{rect_width}{Width of rectangles (used to display observed variables),
Default: 1.2}

\item{rect_height}{Height of rectangles (used to display observed variables),
Default: 0.8}

\item{ellipses_a}{Width of ellipses (used to display latent variables),
Default: 1}

\item{ellipses_b}{Height of ellipses (used to display latent variables),
Default: 1}

\item{spacing_x}{Spacing between columns of the graph, Default: 1}

\item{spacing_y}{Spacing between rows of the graph, Default: 1}

\item{text_size}{Point size of text, Default: 4}

\item{curvature}{Curvature of curved connectors. To flip connectors, use
negative values. Default: .1}

\item{angle}{Angle used to connect nodes by the top and bottom. Defaults to
NULL, which means Euclidean distance is used to determine the shortest
distance between node sides. A numeric value between 0-180 can be provided,
where 0 means that only nodes with the same x-coordinates are connected
top-to-bottom, and 180 means that all nodes are connected top-to-bottom.}
}
\value{
Object of class 'sem_graph'
}
\description{
Starting with nodes, a layout, and edges, prepare several data
objects that can be rendered into a SEM graph. Using this function allows
users to manually change the default graph specification before plotting it.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
