% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{printResults}
\alias{printResults}
\title{Print results of different types of analyses}
\usage{
printResults(x, keepCols = c("label", "est_sig", "confint"),
  digits = 2, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{keepCols}{Character vector, indicating the columns to retain.
Default: c("label", "est_sig", "confint").}

\item{digits}{Integet. Number of digits to round to when formatting;
Default: 2.}

\item{...}{Other arguments passed to and from other functions.}
}
\value{
data.frame
}
\description{
Takes an object, and prints the results as an APA table.
}
