% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{printResultsTable}
\alias{printResultsTable}
\title{Print Mplus results table formatted for publication}
\usage{
printResultsTable(mplusModel, parameters = "unstandardized",
  keepCols = c("label", "est_sig", "confint"), digits = 2, ...)
}
\arguments{
\item{mplusModel}{An mplusModel object, as returned by \code{readModels}.}

\item{parameters}{A character string corresponding to the name of an element
of the $parameters list in \code{mplusModel}. Usually one of
\code{c("unstandardized", "stdyx.standardized", "stdy.standardized")}.}

\item{keepCols}{A character vector of columns to retain from the results
section. Standard column names are
\code{c("paramHeader", "param", "est", "se", "est_se", "pval")}. Special
column names added by \code{printResultsTable} are:
\code{c("est_sig", "confint", "label")}. These correspond to 1) the "est"
column with significance asterisks appended; 2) a formatted confidence
interval; 3) a label, obtained by concatenating the "paramHeader" and "param"
columns.}

\item{digits}{Number of digits to round to when formatting columns.
\code{c("unstandardized", "stdyx.standardized", "stdy.standardized")}.}

\item{...}{Logical expressions used to filter the rows of results returned.}
}
\value{
A data.frame of formatted Mplus results.
}
\description{
Takes an mplusModel object returned by \code{readModels}, and formats it as
a publication-ready table.
}
\examples{
#Make me!
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{est_sig}},
  \code{\link{param_label}}, \code{\link{rbind_tables}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
