% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pseudo_class.R
\name{pseudo_class_data}
\alias{pseudo_class_data}
\title{Sample class membership based on class membership probabilities}
\usage{
pseudo_class_data(fit, x = NULL, m = 20, output_type = "list")
}
\arguments{
\item{fit}{A fitted mx_mixture model}

\item{x}{A dataset to which to add "class", if NULL, takes data from fit object}

\item{m}{How many datasets to generate. Default is 10.}

\item{output_type}{Either 'list' (default), 'long', or 'class_only'. If 'list', the datasets are returned as a list.
If "long", the datasets are combined into a long dataset containing ".imp" to indicate the separate datasets.
If "class_only", class assignments are returned as a matrix with "m" columns.}
}
\value{
Returns a list of size "m" containing data.frame's with variable "class",
or if output_type is "long", one long format data.frame with variable "class".
If output_type is "class_only" a matrix with "m" columns containing class assignments.
}
\description{
Adds a variable named "class" to "m" datasets. The variable contains the
random class membership assignment based on the probabilities of "fit"
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Not run:
x <- iris[,c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
colnames(x) <- c("SL", "SW", "PL", "PW")
tidySEM::mx_profiles(data = x, classes = 3) -> fit

class_assignments <- pseudo_class_data(fit)
## End(Not)
\dontshow{\}) # examplesIf}
}
\author{
Frank Gootjes
}
