% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pseudo_class.R
\name{pseudo_class_pool}
\alias{pseudo_class_pool}
\title{Pool parameter estimates of fit objects}
\usage{
pseudo_class_pool(fits, df_complete = NULL, ...)
}
\arguments{
\item{fits}{List of fit objects (e.g. MxModel or lavaan)}

\item{df_complete}{The residual degrees of freedom for the model}

\item{...}{arguments passed to specific implementations of pseudo_class_pool}
}
\description{
This S3 method pools parameter estimates of a list of fit objects using Rubin's Rules.
The list of fits objects can be acquired with \code{\link[tidySEM]{pseudo_class_technique}}.
In case the fit objects are not of class MxModel or lavaan, \code{\link[mice]{pool}} is tried (if that package is installed).
}
