% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pseudo_class.R
\name{pseudo_class_technique}
\alias{pseudo_class_technique}
\title{Analyze cluster membership using the pseudo-class technique}
\usage{
pseudo_class_technique(
  fit,
  analysis,
  x = NULL,
  m = 10,
  pool_results = TRUE,
  expose_data = FALSE,
  ...
)
}
\arguments{
\item{fit}{A fitted mx_mixture model}

\item{analysis}{Either an expression to execute on every generated dataset,
or a function that performs the analysis on every generated dataset.}

\item{x}{The corresponding dataset on which the model "fit" was fitted. If NULL (default)
the dataset is taken from "fit".}

\item{m}{The amount of datasets to generate. Default is 10.}

\item{pool_results}{Whether to pool the results of the analyses using \code{\link[mice]{pool}}, default is TRUE.}

\item{expose_data}{Whether the expression explicitly refers to the generated dataset using "data".}

\item{...}{Arguments passed to \code{\link[mice]{pool}}}
}
\value{
If pool_results is FALSE, returns a list of size "m" containing fitted models as returned by "analysis".
If pool_results is TRUE, returns an object of class \code{\link[mice]{mipo}}
}
\description{
Given a fitted tidySEM mixture model "fit", and the corresponding dataset "x",
this functions executes "analysis" on "m" datasets. Each of these datasets
contains a "class" variable representing random class assignment based on
the class probabilities of "fit".
}
\note{
Note that if you provide "x", the function assumes rows of "x" correspond to the same
rows of the dataset on which "fit" is based.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Not run:
x <- iris[,c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
colnames(x) <- c("SL", "SW", "PL", "PW")
tidySEM::mx_profiles(data = x, classes = 3) -> fit

pseudo_class_technique( fit = fit,
                        analysis = lm( SL ~ class ) ) -> pct

summary(pct)


pseudo_class_technique(fit = fit,
                       analysis = lm(Sepal.Length ~ class, data = data),
                       x = iris,
                       m = 10,
                       pool_results = TRUE,
                       expose_data = TRUE) -> pcte

summary(pcte)

pseudo_class_technique(fit = fit,
                       analysis = function(df) lm(Sepal.Length ~ class, data = df),
                       x = iris,
                       m = 10,
                       pool_results = TRUE,
                       analysis_method = "function") -> pct_func

summary(pct_func)

pseudo_class_technique( fit = fit,
                        analysis = nnet::multinom( class ~ SL + SW + PL ) ) -> membership_prediction


## End(Not run)
\dontshow{\}) # examplesIf}
}
\author{
Frank Gootjes
}
