% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{rbind_tables}
\alias{rbind_tables}
\title{Row-binds tables for publication}
\usage{
rbind_tables(table_list)
}
\arguments{
\item{table_list}{A list of tables.}
}
\value{
A table.
}
\description{
Converts tables (data.frames, matrices) to character, and row-binds them,
inserting a label into the first column for each sub-table.
}
\examples{
table_list <- list(
  table_f = data.frame(paramHeader = c("F.BY", "F.BY"), param = c("A", "B")),
  table_g = data.frame(paramHeader = c("G.BY", "G.BY"), param = c("A", "B")))
table_list <- list(
  data.frame(paramHeader = c("F.BY", "F.BY"), param = c("A", "B")),
  data.frame(paramHeader = c("G.BY", "G.BY"), param = c("A", "B")))
rbind_tables(table_list)
}
\seealso{
Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{est_sig}},
  \code{\link{param_label}}, \code{\link{table_results}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
