% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.R
\name{report_columns}
\alias{report_columns}
\title{Change which columns table_results reports}
\usage{
report_columns(x = c("label", "est_sig", "se", "pval", "confint",
  "group", "level"))
}
\arguments{
\item{x}{A character vector with column names for
\code{\link[tidySEM]{table_results}}.}
}
\description{
Sets a global option to control which columns table_results reports when
\code{all = FALSE}.
}
\examples{
report_columns(c("label", "est_sig", "se", "pval", "confint", "group"))
}
\seealso{
Other Reporting tools: \code{\link{conf_int}},
  \code{\link{est_sig}}, \code{\link{table_results}}
}
\author{
Caspar J. van Lissa
}
\concept{Reporting tools}
\keyword{reporting}
