% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-bifactor.R
\name{run_bifactor}
\alias{run_bifactor}
\title{Run a bifactor model}
\usage{
run_bifactor(x, total_variance = c("modelled", "observed"), ...)
}
\arguments{
\item{x}{An object of class  \code{tidy_sem_bifactor}
(generated using \code{\link[tidySEM]{bifactor}}).}

\item{total_variance}{whether to compute total variances with the
model implied or the observed correlation matrix.
Default: "modelled"}

\item{...}{Additional parameters passed to \code{\link{run_lavaan}}.}
}
\value{
A list containing a \code{lavaan} fit object (\code{fit}) and omega
statistics (\code{omega}, \code{omega_subfactor},
\code{omegah}, \code{omegah_subfactor}).
}
\description{
Run a bifactor model using lavaan.
This function relies on \code{\link{run_lavaan}} to run the model.
}
\references{
10.1080/00223891.2015.1089249; 10.1080/00223891.2015.1117928
}
