% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmx_functions.R
\name{run_mx}
\alias{run_mx}
\title{Run an OpenMx model with sensible defaults}
\usage{
run_mx(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Parameters passed on to other functions.}
}
\value{
Returns an \code{\link[OpenMx]{mxModel}} with free parameters updated
to their final values.
}
\description{
Run an OpenMx model with sensible defaults.
}
\examples{
df <- iris[1:3]
names(df) <- paste0("X_", 1:3)
run_mx(measurement(tidy_sem(df), meanstructure = TRUE))
}
