% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motley_functions.R
\name{SB_chisq_Pvalues}
\alias{SB_chisq_Pvalues}
\title{Satorra-Bentler corrected chi-square tests for table}
\usage{
SB_chisq_Pvalues(tableChi_df_scf)
}
\arguments{
\item{tableChi_df_scf}{A table of model chi-square values, degrees of freedom
, and scale correction factors.}
}
\value{
A data.frame of chi-square values, degrees of freedom, and p-values
for chi-square difference tests.
}
\description{
Computes Satorra-Bentler corrected chi-square test for a table of model chi-
square values, degrees of freedom, and scale correction factors.
}
\examples{
df <- data.frame(chi2 = c(23, 44, 65), df = c(78, 74, 70), scf = c(1.02, 1.12, 1.28))
SB_chisq_Pvalues(df)
}
\seealso{
\code{\link{SBChisquare}} for a single chi-square test.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{param_label}}, \code{\link{rbind_tables}},
  \code{\link{table_cor}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
