% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motley_functions.R
\name{SBChisquare}
\alias{SBChisquare}
\title{Satorra-Bentler corrected chi-square test}
\usage{
SBChisquare(Chisq1, df1, scf1, Chisq2, df2, scf2)
}
\arguments{
\item{Chisq1}{Chi square value of model 1.}

\item{df1}{Degrees of freedom of model 1.}

\item{scf1}{Scale correction factor of model 1.}

\item{Chisq2}{Chi square value of model 2.}

\item{df2}{Degrees of freedom of model 2.}

\item{scf2}{Scale correction factor of model 2.}
}
\value{
Named numeric vector with chi-square value, degrees of freedom, and
p-value.
}
\description{
Computes Satorra-Bentler corrected chi-square test.
}
\examples{
df <- data.frame(chi2 = c(23, 44, 65), df = c(78, 74, 70), scf = c(1.02, 1.12, 1.28))
SBChisquare(24, 78, 1.02, 65, 70, 1.28)
}
\seealso{
\code{\link{SB_chisq_Pvalues}} to apply SBChisquare to a table of
model chi-square values.

Other Mplus functions: \code{\link{SB_chisq_Pvalues}},
  \code{\link{conf_int}}, \code{\link{corTable}},
  \code{\link{est_sig}}, \code{\link{param_label}},
  \code{\link{rbind_tables}}, \code{\link{table_results}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
