% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_bootstrap.R
\name{srmr}
\alias{srmr}
\title{Calculate Standardized Root Mean Residual}
\usage{
srmr(x, y)
}
\arguments{
\item{x}{An object for which a method of \code{\link[stats:cor]{stats::cor()}} exists
(e.g, \code{data.frame}).}

\item{y}{An object for which a method of \code{\link[stats:cor]{stats::cor()}} exists
(e.g, \code{data.frame}).}
}
\value{
\code{numeric}
}
\description{
Given two datasets, computes the correlation matrix for both,
and then calculates the standardized root mean residual difference between
these two correlation matrices.
}
\examples{
\dontrun{
if(interactive()){
 srmr(iris[1:2], iris[3:4])
}
}
}
