% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_syntax.R
\name{syntax}
\alias{syntax}
\alias{syntax<-}
\title{Extract syntax from sem_syntax}
\usage{
syntax(x)

syntax(x) <- value
}
\arguments{
\item{x}{Object of class sem_syntax.}

\item{value}{A valid value for \code{syntax(x)}.}
}
\value{
data.frame
}
\description{
Provides access to the \code{syntax} element of a
\code{sem_syntax} object. This can be used to return or assign to the
\code{syntax} element.
}
