% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table_results.R
\name{table_cor}
\alias{table_cor}
\title{Extract correlation tables from mplusModel}
\usage{
table_cor(
  mplusModel,
  parameters = "stdyx.standardized",
  valueColumn = "est_sig",
  digits = 2
)
}
\arguments{
\item{mplusModel}{An mplusModel object, as returned by \code{readModels}.}

\item{parameters}{A character string corresponding to the name of an element
of the $parameters list in \code{mplusModel}. Usually one of
\code{c("unstandardized", "stdyx.standardized", "stdy.standardized")}.}

\item{valueColumn}{Character. Which column to use to propagate the matrix.
Defaults to "est_sig", the estimate with significance asterisks.}

\item{digits}{Number of digits to round to when formatting values.}
}
\value{
A Matrix or a list of matrices (in case there are between/within
correlation matrices).
}
\description{
Takes an mplusModel object returned by \code{readModels}, and extracts a
publication-ready correlation matrix.
}
\examples{
#Make me!
}
\seealso{
\code{\link[MplusAutomation]{readModels}}.

Other Mplus functions: 
\code{\link{SBChisquare}()},
\code{\link{SB_chisq_Pvalues}()},
\code{\link{param_label}()},
\code{\link{rbind_tables}()}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
