% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_syntax.R
\name{as.lavaan}
\alias{as.lavaan}
\title{Convert sem_syntax to lavaan syntax}
\usage{
as.lavaan(x, ...)
}
\arguments{
\item{x}{An object of class \code{sem_syntax}}

\item{...}{Additional parameters to be passed to and from functions.}
}
\value{
Character vector.
}
\description{
Final stage in the tidySEM workflow for syntax generation:
Convert the sem_syntax object to lavaan syntax.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
