% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{conf_int}
\alias{conf_int}
\title{Add confidence intervals to Mplus output}
\usage{
conf_int(mplusresults, digits = 2)
}
\arguments{
\item{mplusresults}{An mplusModel object, as returned by \code{readModels}.}

\item{digits}{Integer. The number of digits to round the confidence interval
to.}
}
\value{
A character vector of formatted confidence intervals.
}
\description{
Takes an mplusModel object returned by \code{readModels}, and constructs
nicely formatted confidence intervals. The method depends on the output;
if the results section already contains confidence/credible intervals, these
are reported. If not, the confidence interval is constructed from the "est"
and "se" columns.
}
\examples{
data <- data.frame(est = c(1.325, 2.432), se = c(.05336, .00325))
conf_int(data)
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{corTable}},
  \code{\link{est_sig}}, \code{\link{param_label}},
  \code{\link{rbind_tables}}, \code{\link{table_results}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
