% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.R
\name{conf_int}
\alias{conf_int}
\title{Format confidence intervals to object}
\usage{
conf_int(x, digits = 2, se = NULL, lb = NULL, ub = NULL, ci = 95)
}
\arguments{
\item{x}{A model object for which a method exists.}

\item{digits}{Integer. The number of digits to round the condidence
boundaries to.}

\item{se}{Optional, numeric. Standard error of the parameters.}

\item{lb}{Optional, numeric. Lower boundary of confidence intervals.}

\item{ub}{Optional, numeric. Upper boundary of confidence intervals.}

\item{ci}{Optional, numeric. What percentage CI to use (only used when
computing CI from \code{se}, based on a normal distribution).}
}
\value{
A character vector of formatted confidence intervals.
}
\description{
Takes a model object, and creates APA-formated confidence intervals.
}
\examples{
conf_int(x = c(1.325, 2.432), se = c(.05336, .00325))
}
\seealso{
table_results est_sig

Other Reporting tools: 
\code{\link{est_sig}()},
\code{\link{report_columns}()},
\code{\link{table_results}()}
}
\author{
Caspar J. van Lissa
}
\concept{Reporting tools}
