% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plot_sem.R
\name{edit_graph}
\alias{edit_graph}
\title{Edit graph elements}
\usage{
edit_graph(x, expr, element = "edges", ...)
}
\arguments{
\item{x}{An object of class \code{sem_graph}.}

\item{expr}{expression to evaluate.}

\item{element}{Character. The element of the \code{sem_graph} to edit,
defaults to \code{"edges"}.}

\item{...}{Arguments passed on to \code{\link{within}}.}
}
\value{
An object of class \code{sem_graph}.
}
\description{
Evaluate an R expression within the environment of the elements
of a \code{sem_graph} object, and return the modified \code{sem_graph}.
}
\examples{
p <- prepare_graph(layout = get_layout("x", rows = 1))
p <- edit_graph(p, {colour = "blue"}, element = "nodes")
plot(p)
}
