% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{est_sig}
\alias{est_sig}
\title{Add significance asterisks to object}
\usage{
est_sig(x, digits = 2, sig = NULL)
}
\arguments{
\item{x}{A model object for which a method exists.}

\item{digits}{Integer. The number of digits to round the estimate column to.}

\item{sig}{Optional, a vector of p-values for the default method.}
}
\value{
A character vector of formatted estimates.
}
\description{
Takes a model object, and adds significance asterisks to the estimate column.
}
\examples{
est_sig(c(.222, .3333), sig = c(.054, .045))
}
\seealso{
table_results

Other Reporting tools: \code{\link{conf_int}},
  \code{\link{report_columns}}, \code{\link{table_results}}
}
\author{
Caspar J. van Lissa
}
\concept{Reporting tools}
