% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{est_sig}
\alias{est_sig}
\title{Add significance asterisks to Mplus output}
\usage{
est_sig(mplusresults, digits)
}
\arguments{
\item{mplusresults}{An mplusModel object, as returned by \code{readModels}.}

\item{digits}{Integer. The number of digits to round the "est" column to.}
}
\value{
A character vector of formatted Mplus estimates.
}
\description{
Takes an mplusModel object returned by \code{readModels}, and adds
significance asterisks to the "est" column, based on the "pval" column.
}
\examples{
#Make me!
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{param_label}},
  \code{\link{printResultsTable}},
  \code{\link{rbind_tables}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
