% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motley_functions.R
\name{format_numeric}
\alias{format_numeric}
\title{Format numeric columns}
\usage{
format_numeric(x, digits)
}
\arguments{
\item{x}{A data.frame.}

\item{digits}{The desired number of digits.}
}
\description{
Formats the numeric columns of a data.frame, to round to a specific number
of digits.
}
\examples{
dat <- mtcars
format_numeric(dat, 1)
}
\author{
Caspar J. van Lissa
}
