% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-attributes.R
\name{get_data}
\alias{get_data}
\alias{get_data<-}
\title{Extract data from tidy_sem}
\usage{
get_data(x)

get_data(x) <- value
}
\arguments{
\item{x}{Object of class tidy_sem.}

\item{value}{A valid value for \code{get_data(x)}.}
}
\value{
data.frame
}
\description{
Provides access to the \code{data} element of a
\code{tidy_sem} object. This can be used to return or assign to the
\code{data} element.
}
\examples{
dict <- tidy_sem(iris, split = "\\\\.")
get_data(dict)
}
