% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{get_layout}
\alias{get_layout}
\title{Generate graph layout from a matrix}
\usage{
get_layout(mat = read.table("clipboard", sep = "\\t", stringsAsFactors =
  FALSE))
}
\arguments{
\item{mat}{Matrix. The default reads a selected matrix from the clipboard.
To use this functionality, specify your layout in a spreadsheet program,
select the block of cells, and copy it to the clipboard.}
}
\value{
Object of class 'tidy_layout'
}
\description{
Starting with a matrix, generate a tidy_layout for a SEM graph.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{tidy_graph}
