% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plot_sem.R
\name{get_nodes}
\alias{get_nodes}
\title{Extract nodes from a SEM model object}
\usage{
get_nodes(x, label = paste(name, est_sig, sep = "\\n"), ...)
}
\arguments{
\item{x}{A model object of class \code{mplusObject} or \code{lavaan}.}

\item{label}{Either a character, indicating which column to use for node
labels, or an expression. See Details.
Defaults to \code{paste(name, est_sig, sep = "\n"},
which gives the node name followed by the estimated value with
significance asterisks.}

\item{...}{Additional parameters to be passed to and from other functions.}
}
\value{
An object of class 'tidy_nodes'
}
\description{
Attempts to extract nodes from a SEM model object, where nodes
are defined as observed or latent variables.
}
\details{
When an expression is passed to \code{label}, it is evaluated using
\code{\link{with}}, in the context of a \code{data.frame} containing the
results
of a call to \code{\link{table_results}} on the \code{x} argument, with an
additional column labeled \code{name}, which contains the node names.
}
\examples{
library(lavaan)
res <- sem("dist ~ speed", cars)
get_nodes(res)
}
\keyword{tidy_graph}
