% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sem.R
\name{layout}
\alias{layout}
\title{Generate graph layout}
\usage{
layout(..., rows = NULL)
}
\arguments{
\item{...}{Character arguments corresponding to layout elements. Use node
names, empty strings (""), or NA values.}

\item{rows}{Numeric, indicating the number of rows of the graph.}
}
\value{
Object of class 'tidy_layout'
}
\description{
Generate a tidy_layout for a SEM graph by specifying node names,
and empty strings or \code{NA} values for spaces.
}
\examples{
layout("c", "",  "d",
       "",  "e", "", rows = 2)
}
\seealso{
get_layout
}
\keyword{tidy_graph}
