% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-cfa.R
\name{measurement}
\alias{measurement}
\title{Generate syntax for a measurement model}
\usage{
measurement(x, ...)
}
\arguments{
\item{x}{An object for which a method exists, including \code{tidy_sem}
(generated using \code{\link[tidySEM]{dictionary}}, or \code{data.frame} (for
which \code{\link[tidySEM]{dictionary}} will be run first).}

\item{...}{Additional parameters passed to \code{\link{add_paths}}.}
}
\value{
An object of class \code{tidy_sem}.
}
\description{
Generate syntax for a measurement model for latent variables.
This function relies on \code{\link{add_paths}} to generate syntax.
}
\examples{
dict <- tidy_sem(c("bfi_1", "bfi_2", "bfi_3", "bfi_4", "bfi_5"))
measurement(dict)
}
