% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_simple_thresholds.R
\name{mx_data_quantiles}
\alias{mx_data_quantiles}
\title{Data Quantiles}
\usage{
mx_data_quantiles(df)
}
\arguments{
\item{df}{A \code{data.frame} with only columns of class \code{ordered}.}
}
\value{
Matrix
}
\description{
Get quantiles based on empirical normal distribution of data.
}
