% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-attributes.R
\name{nodes}
\alias{nodes}
\alias{nodes<-}
\title{Extract nodes from sem_graph}
\usage{
nodes(x)

nodes(x) <- value
}
\arguments{
\item{x}{Object of class sem_graph.}

\item{value}{A valid value for \code{nodes(x)}.}
}
\value{
data.frame
}
\description{
Provides access to the \code{nodes} element of a
\code{sem_graph} object. This can be used to return or assign to the
\code{nodes} element.
}
\examples{
edg <- data.frame(from = "x", to = "y")
p <- prepare_graph(edges = edg, layout = get_layout("x", "y", rows = 1))
nodes(p)
}
