% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_bootstrap.R
\name{pmc}
\alias{pmc}
\alias{pmc_srmr}
\title{Predictive Model Comparison}
\usage{
pmc(x, ..., reps = 20, ci = 0.95, FUN = NULL)

pmc_srmr(x, ..., reps = 20, ci = 0.95)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Additional arguments.}

\item{reps}{Number of data sets to simulate, Default: \code{100}.}

\item{ci}{Confidence interval coverage, Default: \code{0.95}.}

\item{FUN}{Function used to compare the real data (referred to as \code{x}) to the
model-implied data (referred to as \code{y}). Defaults to \code{NULL}, which uses
\code{\link[=chi_sq]{chi_sq()}} for models with all ordinal variables, and
\code{\link[=srmr]{srmr()}} otherwise, treating all variables as continuous.}
}
\value{
A \code{data.frame}.
}
\description{
Compare (non-nested) models estimated on the same data using
model-simulated data.
}
