% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_bootstrap.R
\name{pmc_srmr}
\alias{pmc_srmr}
\title{Predictive Model Comparison}
\usage{
pmc_srmr(x, ..., reps = 20, ci = 0.95)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Additional arguments.}

\item{reps}{Number of data sets to simulate, Default: \code{100}.}

\item{ci}{Confidence interval coverage, Default: \code{0.95}.}
}
\value{
A \code{data.frame}.
}
\description{
Compare (non-nested) models estimated on the same data using
model-simulated data.
}
