% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pseudo_class.R
\name{pseudo_class_technique}
\alias{pseudo_class_technique}
\title{Analyze cluster membership using the pseudo-class technique}
\usage{
pseudo_class_technique(
  fit,
  analysis,
  x = NULL,
  m = 20,
  pool_results = FALSE,
  expose_data = detect_usage_of_data(substitute(analysis)),
  ...
)
}
\arguments{
\item{fit}{A fitted mx_mixture model}

\item{analysis}{Either a call to execute on every generated dataset,
or a function that performs the analysis on every generated dataset,
or a character that can be converted with \code{\link[tidySEM]{as_ram}}.}

\item{x}{The corresponding dataset on which the model "fit" was fitted. If NULL (default)
the dataset is taken from "fit".}

\item{m}{The amount of datasets to generate. Default is 10.}

\item{pool_results}{Whether to pool the results of the analyses using \code{\link[tidySEM]{pseudo_class_pool}}, default is FALSE.}

\item{expose_data}{Whether the expression explicitly refers to the generated dataset using "data".
The default is to detect the usage of the "data" keyword in 'analysis'.}

\item{...}{Arguments passed to \code{\link[tidySEM]{pseudo_class_pool}}}
}
\value{
If pool_results is FALSE, returns a list of size "m" containing fitted models as returned by "analysis".
If pool_results is TRUE, returns a data frame containing pooled results.
}
\description{
Given a fitted tidySEM mixture model "fit", and the corresponding dataset "x",
this functions executes "analysis" on "m" datasets. Each of these datasets
contains a "class" variable representing random class assignment based on
the class probabilities of "fit".
}
\note{
Note that if you provide "x", the function assumes rows of "x" correspond to the same
rows of the dataset on which "fit" is based.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Not run:
x <- iris[,c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
colnames(x) <- c("SL", "SW", "PL", "PW")
tidySEM::mx_profiles(data = x, classes = 3) -> fit

pseudo_class_technique( fit = fit,
                        analysis = "SL ~ class",
                        pool_results = TRUE ) -> pct_mx

summary(pct_mx)

pseudo_class_technique( fit = fit,
                        analysis = lm( SL ~ class ),
                        pool_results = TRUE  ) -> pct_lm

summary(pct_lm)


pseudo_class_technique(fit = fit,
                       analysis = lm(Sepal.Length ~ class, data = data),
                       x = iris,
                       m = 10,
                       pool_results = TRUE,
                       expose_data = TRUE) -> pcte

summary(pcte)

pseudo_class_technique(fit = fit,
                       analysis = function(df) lm(Sepal.Length ~ class, data = df),
                       x = iris,
                       m = 10,
                       pool_results = TRUE,
                       analysis_method = "function") -> pct_func

summary(pct_func)


## End(Not run)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Paper recommending, among others, the pseudo-class technique.
The GRoLTS-Checklist: Guidelines for Reporting on Latent Trajectory Studies. https://doi.org/10.1080/10705511.2016.1247646
\item Paper describing the pseudo-class technique.
Residual Diagnostics for Growth Mixture Models. https://doi.org/10.1198/016214505000000501
\item Paper describing pooling results.
Applied missing data analysis with SPSS and (R) Studio. Chapter Rubin's Rules. https://bookdown.org/mwheymans/bookmi/rubins-rules.html
\item Source for the default of 20 draws.
Wald Test of Mean Equality for Potential Latent Class Predictors in Mixture Modeling. https://www.statmodel.com/download/MeanTest1.pdf
}
}
\author{
Frank Gootjes
}
