% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax-attributes.R
\name{syntax}
\alias{syntax}
\alias{syntax<-}
\title{Extract syntax from tidy_sem}
\usage{
syntax(x)

syntax(x) <- value
}
\arguments{
\item{x}{Object of class tidy_sem.}

\item{value}{A valid value for \code{syntax(x)}.}
}
\value{
data.frame
}
\description{
Provides access to the \code{syntax} element of a
\code{tidy_sem} object. This can be used to return or assign to the
\code{syntax} element.
}
\examples{
dict <- tidy_sem(iris, split = "\\\\.")
dict <- add_paths(dict, Sepal.Width ~~ Sepal.Length)
syntax(dict)
}
