% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-mx_ordinal.R
\name{table_prob}
\alias{table_prob}
\title{Results table in probability scale}
\usage{
table_prob(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Arguments passed to other functions.}
}
\value{
A data.frame with results in probability scale.
}
\description{
Returns thresholds for ordinal
dependent variables in probability scale.
}
\examples{
\dontrun{
df <- data_mix_ordinal
df[1:4] <- lapply(df, ordered)
mx_lca(data = df,
       classes = 2) -> res
}
}
\seealso{
Other Reporting tools: 
\code{\link{conf_int}()},
\code{\link{est_sig}()},
\code{\link{table_fit}()},
\code{\link{table_results}()}
}
\concept{Reporting tools}
\keyword{reporting}
