% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{table_results}
\alias{table_results}
\title{Print results table formatted for publication}
\usage{
table_results(x, standardized = TRUE, columns = c("label", "est_sig",
  "se", "pval", "ci"), digits = 2, ...)
}
\arguments{
\item{x}{A model object for which a method exists.}

\item{standardized}{Logical. Whether to return standardized parameters or
not. Defaults to TRUE.}

\item{columns}{A character vector of columns to retain from the results
section. Defaults to \code{c("label", "est_sig", "se", "pval", "ci")}. These
correspond to 1)
the parameter label, 2) estimate column with significance asterisks appended
(\* <.05, \*\* < .01, \*\*\* < .001); 3) standard error, 4) p-value, 5) a
formatted confidence interval.}

\item{digits}{Number of digits to round to when formatting numeric columns.}

\item{...}{Logical expressions used to filter the rows of results returned.}
}
\value{
A data.frame of formatted Mplus results.
}
\description{
Takes a model object, and formats it as a publication-ready table.
}
\examples{
#Make me!
}
\seealso{
\code{\link{readModels}}.

Other Mplus functions: \code{\link{SBChisquare}},
  \code{\link{SB_chisq_Pvalues}}, \code{\link{conf_int}},
  \code{\link{corTable}}, \code{\link{est_sig}},
  \code{\link{param_label}}, \code{\link{rbind_tables}}
}
\author{
Caspar J. van Lissa
}
\concept{Mplus functions}
