% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.R
\name{table_results}
\alias{table_results}
\title{Print results table formatted for publication}
\usage{
table_results(x, standardized = TRUE, all = FALSE, digits = 2, ...)
}
\arguments{
\item{x}{A model object for which a method exists.}

\item{standardized}{Logical. Return standardized parameters, or not.
Defaults to TRUE.}

\item{all}{Logical. Whether to return all available results columns
(including columns generated by \code{table_results}, or whether to return a
simplified table containing the essential columns for APA-style reporting.
Defaults to FALSE. If set to TRUE, you can manually select the desired
columns.
formatted confidence interval.}

\item{digits}{Number of digits to round to when formatting numeric columns.}

\item{...}{Logical expressions used to filter the rows of results returned.}
}
\value{
A data.frame of formatted Mplus results.
}
\description{
Takes a model object, and formats it as a publication-ready table.
}
\examples{
#Make me!
}
\seealso{
Other Reporting tools: \code{\link{conf_int}},
  \code{\link{est_sig}}, \code{\link{report_columns}}
}
\author{
Caspar J. van Lissa
}
\concept{Reporting tools}
\keyword{reporting}
